torch 1.0, simplify data_loader, fix data processing, fix tests

+++ tests/datasets/test_hilary.py
@@ -8,7 +8,6 @@ from src.datasets import hilary_dataset
 from src.utils import Checkpoint
 
 from tests.datasets.utils import _download_file_from_drive_side_effect
-from tests.datasets.utils import compute_spectrogram_side_effect
 
 data_directory = Path('tests/_test_data/')
 
@@ -18,16 +17,15 @@ def cleanup():
     yield
     cleanup_dir = data_directory / 'Hilary'
     print("Clean up: removing {}".format(cleanup_dir))
-    shutil.rmtree(str(cleanup_dir))
+    if cleanup_dir.is_dir():
+        shutil.rmtree(str(cleanup_dir))
 
 
 @mock.patch("src.utils.Checkpoint.from_path")
-@mock.patch("src.datasets.hilary.compute_spectrogram")
 @mock.patch("torchnlp.download._download_file_from_drive")
 @pytest.mark.usefixtures("cleanup")
-def test_hilary_dataset(mock_download_file_from_drive, mock_compute_spectrogram, mock_from_path):
+def test_hilary_dataset(mock_download_file_from_drive, mock_from_path):
     mock_download_file_from_drive.side_effect = _download_file_from_drive_side_effect
-    mock_compute_spectrogram.side_effect = compute_spectrogram_side_effect
     mock_from_path.return_value = Checkpoint(directory='.', model=lambda x: x, step=0)
 
     # Check a row are parsed correctly
