+++ tests/datasets/test_lj_speech.py
@@ -9,7 +9,6 @@ from src.datasets import lj_speech_dataset
 from src.utils import Checkpoint
 
 from tests.datasets.utils import urlretrieve_side_effect
-from tests.datasets.utils import compute_spectrogram_side_effect
 
 lj_directory = Path('tests/_test_data/')
 
@@ -48,16 +47,15 @@ def cleanup():
     yield
     cleanup_dir = lj_directory / 'LJSpeech-1.1'
     print("Clean up: removing {}".format(cleanup_dir))
-    shutil.rmtree(str(cleanup_dir))
+    if cleanup_dir.is_dir():
+        shutil.rmtree(str(cleanup_dir))
 
 
 @mock.patch("src.utils.Checkpoint.from_path")
-@mock.patch("src.datasets.lj_speech.compute_spectrogram")
 @mock.patch("urllib.request.urlretrieve")
 @pytest.mark.usefixtures("cleanup")
-def test_lj_speech_dataset(mock_urlretrieve, mock_compute_spectrogram, mock_from_path):
+def test_lj_speech_dataset(mock_urlretrieve, mock_from_path):
     mock_urlretrieve.side_effect = urlretrieve_side_effect
-    mock_compute_spectrogram.side_effect = compute_spectrogram_side_effect
     mock_from_path.return_value = Checkpoint(directory='.', model=lambda x: x, step=0)
 
     # Check a row are parsed correctly
@@ -74,19 +72,18 @@ def test_lj_speech_dataset(mock_urlretrieve, mock_compute_spectrogram, mock_from
     assert len(dev) == 13100 * 0.2
 
     # Check sum to ensure its the same exact split
-    assert sum([len(r['text']) for r in dev]) == 258045
-    assert sum([len(r['text']) for r in train]) == 1052287
+    assert sum([len(r.text) for r in dev]) == 258045
+    assert sum([len(r.text) for r in train]) == 1052287
 
     # Test deterministic shuffle
-    assert train[0]['text'] == (
+    assert train[0].text == (
         'Once a warrant-holder sent down a clerk to view certain goods, and the clerk found that '
         'these goods had already a "stop" upon them, or were pledged.')
-    assert 'tests/_test_data/LJSpeech-1.1/wavs/LJ014-0331.wav' in str(
-        train[0]['aligned_audio_path'])
-    assert dev[0]['text'] == (
+    assert 'tests/_test_data/LJSpeech-1.1/wavs/LJ014-0331.wav' in str(train[0].audio_path)
+    assert dev[0].text == (
         'Mister Mullay went, and a second interview was agreed upon, when a third person, '
         'Mister Owen,')
-    assert 'tests/_test_data/LJSpeech-1.1/wavs/LJ011-0243.wav' in str(dev[0]['aligned_audio_path'])
+    assert 'tests/_test_data/LJSpeech-1.1/wavs/LJ011-0243.wav' in str(dev[0].audio_path)
 
     _re_filename = re.compile('LJ[0-9]{3}-[0-9]{4}')
 
@@ -94,9 +91,9 @@ def test_lj_speech_dataset(mock_urlretrieve, mock_compute_spectrogram, mock_from
     seen = 0
     for data in [train, dev]:
         for row in data:
-            basename = row['aligned_audio_path'].name[:10]
+            basename = row.audio_path.name[:10]
             assert _re_filename.match(basename)
             if basename in verbalize_test_cases:
                 seen += 1
-                assert verbalize_test_cases[basename] in row['text']
+                assert verbalize_test_cases[basename] in row.text
     assert seen == len(verbalize_test_cases)
